+++ libs/dot-components/src/lib/components/checkbox/Checkbox.tsx
@@ -14,13 +14,22 @@ export type CheckboxSize = 'medium' | 'small';
 export type CheckboxLabelPlacement = 'bottom' | 'end' | 'start';
 
 export interface CheckboxProps extends RadioButtonProps {
+  /** accessibility label */
+  ariaLabel?: string;
+  /** accessibility labelled by */
+  ariaLabelledby?: string;
+  /** If true, the ripple effect will be disabled. */
+  disableRipple?: boolean;
   /** if true the checkbox will display with intermediate */
   indeterminate?: boolean;
 }
 
 export function DotCheckbox({
+  ariaLabel,
+  ariaLabelledby,
   checked,
   className,
+  disableRipple,
   'data-testid': dataTestId,
   disabled,
   id,
@@ -51,8 +60,13 @@ export function DotCheckbox({
           color="primary"
           data-testid={dataTestId}
           disabled={disabled}
+          disableRipple={disableRipple}
           id={id}
           indeterminate={indeterminate}
+          inputProps={{
+            'aria-label': ariaLabel ? ariaLabel : label,
+            'aria-labelledby': ariaLabelledby,
+          }}
           inputRef={inputRef}
           name={name}
           onChange={handleChange}